backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
celery[redis]==5.4.0
redis==5.2.0
docker==7.1.0
gitpython==3.1.43
requests==2.31.0
urllib3<2.0
alembic==1.14.0
cryptography==43.0.0
python-multipart==0.0.12
websockets==13.0
pyyaml==6.0.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
